(+) Categories controller added, new features based on the tags
Category controller added, for model management. Button on the navbar
links to the Category creation, instead of Post creation;

Posts are now created with the blue button placed after the page title
OR in the blank page screen, replacing the previous message;

Now, post creation when browsing a certain Category will prompt the form
with that same Category selected by default OR a placeholder message if
the post is created at Home.