(~) Removes redundant code and unnecessary method atributions
- In Post controller, there was a line used to set slug to nil
in order to fix the problem of slugs not updating. Since a more
elegant solution was already defined in Post model, there was no
need to keep this attribution

- In Comment controller, there was a variable @fallback used with
post path storing purpose, later used in redirect_back. This one
was removed and the fallback path is now the root path